[initconfig] Migrate to a single config file
Migrate all configuration in initconfig to <librarydirectory>/as3lib.cfg
Make use configparser functions instead of what was previously done to fetch values. This might be slightly slower but it is much nicer to look at.
Will delete wayland.cfg if present but not mm.cfg as that isn't a file that is unique to this library and the user might still want that.