[interface_tk] Overhaul most things by making custom widgets for easy attribute application
Attributes are now applied the same way on most widgets. This eliminates a lot of if statements and makes things a bit faster.
The custom widgets aren't done yet but I have to rewrite those anyways. The childproperties dictionary will probably be removed once these are done because it won't be needed anymore.