[flash.utils] Fix Timer exectution and implement stopping when setting repeatCount
Timer execution wasn't done properly before. It was waiting for the event to execute before restarting the timer (the timer delay would also include the execution time of the event).
Timer is supposed to stop when repeatCount is set to less than or equal to currentCount. I think I implemented this correctly.